+++ packages/components/src/Chips/index.ts
@@ -1,3 +1,4 @@
 export { ChipDismissible } from "./ChipDismissible";
 export { Chips } from "./Chips";
 export { Chip } from "./Chip";
+export { InternalChipButton } from './InternalChipButton'