+++ packages/site/vite.config.ts
@@ -3,15 +3,14 @@ import react from '@vitejs/plugin-react';
 import mdx from '@mdx-js/rollup';
 import path from 'path';
 
-// https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react(), mdx()],
   optimizeDeps: {
-    include: ['@jobber/formatters']
+    include: ['@jobber/formatters', '@jobber/hooks']
   },
   resolve: {
     alias: {
-      '@jobber/hooks': path.resolve(__dirname, 'packages/hooks/src'),
+      '@jobber/hooks': path.resolve(__dirname, '../hooks/src'),
     },
   },
   server: {